replaces queue#length with #size
length is marked as deprecated.

crystal hash/array standardize on #size so this brings mosquito in line with language convention

+++ src/mosquito/backend.cr
@@ -56,6 +56,6 @@ module Mosquito
     abstract def finish(task : Task) # should this be called succeed?
     abstract def terminate(task : Task) # should this be called fail?
     abstract def flush : Nil
-    abstract def size : Int32
+    abstract def size : Int64
   end
 end
